feat:拣货单打印优化

+++ erp-model/erp-model-wms/src/main/java/com/erp/model/wms/dto/pickingstrategy/PickingListsDTO.java
@@ -245,7 +245,7 @@ public class PickingListsDTO {
         /**
          * 父级sku数量
          */
-        private Double parentSkuQty;
+        private Integer parentSkuQty;
 
         /**
          * 子级sku
